Introduce DoxygenProcessHandler and AutoProjectInfo
We want to separate out all the ugly subprocess call and stuff into a
separate helper class to keep it as clean as possible. We inject it with
the various bits it needs to get the job done.

We create a AutoProjectInfo concept which handled separately and stored
separately from the standard ProjectInfo. This acts as a temporary for
the parts of the code which need to think in terms of source code and
generating the doxygen output and then we use it to create a standard
ProjectInfo instance and pass that on to the rest of the code. It
separates out the concepts and allows us to keep the ProjectInfo class
immutable.

We also introduce the TransformWrapper class to get around the fact that
the Sphinx and Docutils APIs expect you to pass a class object directly
to them so they can take care of instantiation. This prevents easy
dependency injection which is annoying so we create a wrapper with a
__call__ method which is going to pretend to be the class object but
allows us to inject the helper classes we need.